fix: improvements to onboarding (#65)
A bunch of improvements and fixes related to the getting started
experience with draft. Thanks to @josephhilby for helping find some of
these.

Changes (in no particular order):
- Removed old `.vscode/launch.json` configurations (and fixed others)
- Added a Docker compose file for easily starting core services (but see
#69 for an even better way to do this)
- Removed old (unused) deployment files
- Cleaned up logging and configuration for blueprint
- Removed unused catalyst web-client
- Added catalyst config file and updated chassis dependency
- Updated example services to use latest chassis and updated configs to
match new core service deployment
- Removed unnecessary expose of ports in Dockerfile
- Added infra configs to dctl context
- Added a `--full` flag to `infra clean` to allow removing contains
without deleting config and data

+++ tools/dctl/cmd/infra/clean.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"strings"
 
 	"github.com/steady-bytes/draft/tools/dctl/docker"
 	"github.com/steady-bytes/draft/tools/dctl/output"
@@ -22,46 +23,54 @@ func Clean(cmd *cobra.Command, args []string) (err error) {
 		config, err := getInfraConfig(name)
 		if err != nil {
 			output.Error(err)
-			return err
+			continue
 		}
 
-		err = dockerCtl.RemoveContainerByName(ctx, containerName(name))
-		if err != nil {
+		err = dockerCtl.StopContainerByName(ctx, containerName(name))
+		if err != nil && !strings.Contains(err.Error(), "not found") {
 			output.Error(err)
 		}
 
-		if config.configFile != nil {
-			output.Print("Deleting configuration file for: %s", name)
-			home, err := os.UserHomeDir()
-			if err != nil {
-				output.Error(err)
-				os.Exit(1)
+		err = dockerCtl.RemoveContainerByName(ctx, containerName(name))
+		if err != nil && !strings.Contains(err.Error(), "not found") {
+			output.Error(err)
+		}
+
+		if Full {
+			if config.configFile != nil {
+				output.Print("Deleting configuration file for: %s", name)
+				home, err := os.UserHomeDir()
+				if err != nil {
+					output.Error(err)
+					continue
+				}
+
+				dirName := filepath.Join(home, ".config", "dctl", "infra")
+				fileName := filepath.Join(dirName, fmt.Sprintf("%s.yaml", name))
+				err = os.Remove(fileName)
+				if err != nil && !os.IsNotExist(err) {
+					output.Error(err)
+					continue
+				}
 			}
 
-			dirName := filepath.Join(home, ".config", "dctl", "infra")
-			fileName := filepath.Join(dirName, fmt.Sprintf("%s.yaml", name))
-			err = os.Remove(fileName)
-			if err != nil {
-				output.Error(err)
-				os.Exit(1)
+			if config.mountPoint != nil {
+				output.Print("Deleting volume for: %s", name)
+				home, err := os.UserHomeDir()
+				if err != nil {
+					output.Error(err)
+					continue
+				}
+
+				dirName := filepath.Join(home, ".config", "dctl", "infra", name)
+				err = os.RemoveAll(dirName)
+				if err != nil {
+					output.Error(err)
+					continue
+				}
 			}
 		}
 
-		if config.mountPoint != nil {
-			output.Print("Deleting volume for: %s", name)
-			home, err := os.UserHomeDir()
-			if err != nil {
-				output.Error(err)
-				os.Exit(1)
-			}
-
-			dirName := filepath.Join(home, ".config", "dctl", "infra", name)
-			err = os.RemoveAll(dirName)
-			if err != nil {
-				output.Error(err)
-				os.Exit(1)
-			}
-		}
 	}
 
 	return nil
